Add icons to index files
With URLs that may include any number of directory levels before
even the prefix of our upload, it's difficult (though not impossible)
to upload the icons to a fixed location in the container and
reference that location.  A more self-contained approach is just
to embed the icon data directly into the HTML.  That is what is
done here.

Change-Id: I12342aa479bac41eb3b401d1e92689a56b3c2a2b