Bump ansible test version to 2.5
Iab951c11b23a24757cf5334b36bc8f7d12e19db0 was merged in June to update
Zuul to Ansible 2.5.  This is outside the 4 week update period for new
features, so we can update and use Ansible 2.5 features in zuul-jobs.

Add a little bit more of a note in the requirements about what's going
on while here.

Change-Id: Ic018b7259749148660e1bdb6c9a779551390fd09