upload-logs-swift: Turn FileList into a context manager
The FileList is dynamic and needs to be able to keep track of things
to cleanup.  For example, it has index files added to it from
temporary files which should be removed up when we're finished with
the list.  In a future change we propose a similar addition of a
download script for logs which should also be managed.

Turn the FileList into a context manager.  Modify the index generation
to not create a new FileList, but just replace the internal list.  Use
this for the life-span of the upload by wrapping the relevant parts in
a "with:" statement.

Change-Id: I7135bf5a55d133ce146e9aa84f00041fc8125cbc