Install docker from distro properly
The install-docker role was installing the wrong package name for distro
package based installations. docker-engine is ancient on debuntu and
appears to be called docker.io on both debian and ubuntu supported
releases now. Go ahead and install docker.io on the Debian family of
distros and docker on everything else (fedora, centos, suse, etc).

Change-Id: I0d7a9b8257bd6db86bc6e3d8bb3d6c3c021d7954