Update ANSIBLE_LIBRARY to use envsitepackagesdir
We install zuul via test-requirements, so the zuul files should be in a
directory under the site-packges of the virtualenv that tox installed it
in to. Update the path to point correctly to that location.

Remove the ansible-lint skip tags which should now work because the
library path should be pointing to a location that actually holds the
content.

Change-Id: If2d4b39267c4b9a3102a951143b568f8447af8d9