mirror-workspace-git: use port when pushing git repo through ssh
In the case of using a custom SSH port to connect to the ansible host,
the task of pushing the repository failed, as it assumed the SSH port to
be the default 22.

Now the command is aware of the custom SSH port, as it is specified in
the host facts.

Change-Id: I3418016ad5a57e3024a8a553340a7e69b7d1c2fe