mirror-workspace-git: use port when pushing git repo through ssh
In the case of using a custom SSH port to connect to the ansible host,
the task of pushing the repository failed, as it assumed the SSH port to
be the default 22.

Now the command is aware of the custom SSH port, as it is specified in
the host facts.

Note that we have to escape the hostname if it is an ipv6 address
(because : is a special character in ipv6 addresses). Thankfully ansible
provides a filter to do that for us:
https://docs.ansible.com/ansible/2.5/user_guide/playbooks_filters_ipaddr.html#wrapping-ipv6-addresses-in-brackets

Change-Id: I76969850fc3a57078370fe38efdebac146720797