mirror-workspace-git-repos: Explicitly show HEAD of checked out branches
This expilicitly notes the HEAD of the checked out tree in the logs.
Currently you'll see something like

 ...  HEAD is now at 9dfb46a18 [wip] install ansible as editable during devel jobs
 ... Already on 'master'
 ... ok: Item: Runtime: 0:00:00.079626

for each project as this loop happens.  However there's no explicit
mention of the project name and you have to reverse engineer what
menioned change relates to what project.

Where this would be useful is when you have an external project
dependency and suddenly your testing stops working due to a change in
it.  You can quickly compare the last good run with the failing HEAD
mentioned here to help bisect the problem.

Change-Id: I84d9a8a460e99d3bc558b3a7584d787b22f2ed33