Vendor the RDO repository configuration for installing OVS
Installing OVS on CentOS requires one RDO repository and the
centos-release-openstack packages sets up two additional repos which
aren't required.

In addition, we removed the centos-release-openstack package but this
did not remove the two extra repositories, causing them to "leak" into
the jobs.

This sets up the one repository we need and ensures it is removed after
OVS is installed.

Change-Id: Ida2299a4356282d23b79fac6753b6171211a7651