Pin nodejs package to upstream repo
In the case where a node_version is less than what's in the distro,
we'll wind up installing distro packages instead of upstream packages
which will either fail due to binary path naming issues, or succeed but
install a different version of nodejs than was requested.

Pin the package to the nodesource repo.

Put the file in as a template, as we should really update this role to
be able to optionally make use of a local mirror, in which case the file
will need to jinja.

Change-Id: If313d895a73bd2be7f598179a51b6485a37000ba