Add role to move docs and artifacts to log root
When we're operating in the context of a change, we'd like to
upload artifacts and docs to the log server so that people can iterate
on them before publication safely. Add a role which is intended to be
run after fetch-output that moves the content on the executor so that
normal log upload concepts can be applied.

Change-Id: I1791d1283a0c1f112958db8861526145abe1357c