Fix upload-puppetforge role
The role failed because the puppet-blacksmith gem required
ruby-dev, gcc/g++ compiler and make to compile the gems that
had native extensions.

Fixes the Rakefile since the puppet-blacksmith module
requires a "build" rake task, we can just empty that out
since this role assumed the pkg/ folder in the module
(zuul project dir) already contains the built module.

Setups the GEM_HOME and GEM_BIN folders correctly so that
we install gems in a predictable path which we can then
pass as an environment to the rake command so that it
knows where the actual commands are stored.

Change-Id: I2b39dd69ab56d51e169b7ac6b77c675646006f10