Fix concurrency and syntax issue in merge-output-to-logs
In a multi-node job, these mv commands run in parallel which makes
the second invocation weird and racy (even though there is a mv -f
there) Put in run_once - we don't need to perform this more than
one time, no matter how many nodes there are in the nodeset.

Also, when there are actually files present, if ! $(ls work_dir/*)
winds up expanding as a command to be run.

Change-Id: Ia59c604d6639fab762fae5e53790da8992437587