Fix ownership change in stage-output
If the stage_dir/logs directory was a symlink, Ansible would refuse
to perform the chmod.  However, if we tell Ansible that it is a
directory, it's fine.  This (perhaps curiously) does not tell
Ansible to turn the symlink into a directory.  It just tells it
that everything is okay.

Change-Id: Idbd78b97b95ce50b23dda8f384e32359d31013e3