Load buildset registry data from zuul_return
The push-to-intermediate-registry role is meant to be used in the
post playbook of a job which most likely started the buildset
registry in its pre-playbook.  The buildset_registry variable
is returned to zuul and therefore available to child jobs, but
not to further playbooks in this job.  Therefore, load the variable
back in to memory here.

We may add this functionality directly to Zuul, in which case this
can be removed later.

Change-Id: I81899a9be92789fc1e5858f985eac41871b9d87a