Load buildset_registry in intermediate pull
The intermediate registry pull role is designed to be used in
the same playbook as the run-buildset-registry role, which sets
the buildset_registry fact.  However, that fact is set on the
host where the registry runs, not localhost.  Theoretically we
should be able to delegate setting that fact to all hosts in
the inventory, plus localhost, however, that doesn't seem to
work in local testing.

Work around this by, once again, loading the buildset_registry
fact from the zuul_return file.

Change-Id: Ia16b3af8782c875e64ad5eeeeb5f107482a3e30a