use-buildset-registry: configure as a pull-through proxy
Docker has a curious behavior with respect to the registry-mirror
setting: it will only use the credentials for dockerhub itself.
Therefore, configure the buildset registry to be a pull-through
cache for dockerhub, but with no credentials, so that we can set
buildset registry credentials under the dockerhub url without
their actually being used for dockerhub.

Change-Id: I5daf93868d408d71cf2766fbe1bb17f2cfa8a3c7