run-buildset-container: fix username/password for proxy registry
This configuration is intended to mask the creds supplied by the
local docker configuration when the registry is used as a
pull-through proxy to upstream.  But we need to set the creds
to the empty string, rather than null, which ends up as the string
"None" in the docker config.

Change-Id: Ic36a5c764dec8df4fa460f6735ff5668a1dd3b50