stage-output: fix the archiving of all files
The archival task currently relies on the content of all_sources,
but some files are renamed afterwards if their extensions match
one of the special extension of the list.
This means that few files from all_sources are not found and then
not compressed.
Change then the logic: simply discover all files inside logs
and compress all of them.

Change-Id: I7d34d7d90849736b7b842c0bdd67492816f98ebc